+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 - We added the possibility to configure the email provided to unpaywall. [#14340](https://github.com/JabRef/jabref/pull/14340)
 - We added a "Regenerate" button for the AI chat allowing the user to make the language model reformulate its response to the previous prompt. [#12191](https://github.com/JabRef/jabref/issues/12191)
 - We added support for transliteration of fields to English and automatic transliteration of generated citation key. [#11377](https://github.com/JabRef/jabref/issues/11377)
+- git add jablb/src/test/java/org/jabref/model/entry/DateTest.javaAdded support for BibLaTeX date ranges and improved normalization in the date editor. [#14289](https://github.com/JabRef/jabref/pull/14289)
+
 
 ### Changed